requirements.txt
numpy==1.19.5
opencv_python==3.4.2.17
tensorflow-cpu==2.4.0
streamlit==0.69.0
streamlit_webrtc==0.5.0
imutils